pictobrickWebApp/ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pictobrickwebapp"  # you can rename later to ptb-ml if you want
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    # preprocessing engine deps (Milestone 1)
    "pillow>=10",
    "opencv-python-headless>=4.8",
    "imagehash>=4.3",
    "pydantic>=2,<3",
    "pillow-heif>=1.2.1",
    "pytest>=9.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
